octopoes/pyproject.toml
[tool.poetry]
name = "octopoes"
version = "3.0.0"
description = "KAT's knowledge graph engine"
authors = ["RDO Beheer"]
license = "EUPL"

[build-system]
requires = ["setuptools>=62.2", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta:__legacy__"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.vulture]
min_confidence = 90
paths = ["octopoes", "tests"]

[tool.pylint.format]
max-line-length = "120"

[tool.pylint."MESSAGES CONTROL"]
disable = ["R0914","W1514", "R0903", "E0401", "R0401"] # Ignore too-many-local-variables; unspecified encodings; too-few-public-methods; import-error; cyclic-import
extension-pkg-whitelist = "pydantic" # Ignore pydantic import checks

[tool.black]
line-length = 119
target-version = ['py38']

[tool.pydocstyle]

[tool.mypy]
python_version = "3.8"
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false # Needed for FastAPI decorators

[tool.pytest.ini_options]
addopts = "--cov --cov-branch --cov-report=term-missing:skip-covered"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
pydantic = "^1.10.4"
fastapi = "^0.89.1"
fastapi-utils = "^0.2.1"
uvicorn = "^0.20.0"
pika = "^1.3.1"
PyYAML = "^6.0"
graphql-core = "^3.2.3"
mmh3 = "^3.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
pylint = "^2.15.3"
black = "^22.10.0"
flake8 = "^5.0.4"
vulture = "^2.6"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
robotframework = "^6.0.1"
robotframework-requests = "^0.9.4"
robotframework-tidy = "^3.3.2"
robotframework-httplibrary = "^0.4.2"
pre-commit = "^2.20.0"
